requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-openai
langchain-chroma
chromadb
